Refactor unit tests into separate test files
